judge-service/requirements.txt
torch==2.7.0
loguru==0.7.3
fastapi==0.115.12
uvicorn==0.34.3
pydantic==2.11.4
pydantic-settings==2.7.1
pillow==11.2.1
vllm==0.9.1
transformers==4.52.4